requirements.txt
fastapi==0.116.1
python-dotenv==1.0.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.42
psycopg2-binary==2.9.10
pydantic==2.11.7
httpx==0.28.1
pytest==8.4.1
pytest-asyncio==0.24.0
anyio==4.10.0
PyJWT[crypto]>=2.9.0
bcrypt==5.0.0
python-multipart==0.0.20
slowapi>=0.1.9
email-validator>=2.0.0
